+++ src/lib/sections/app/settings/settings-dialog.svelte
@@ -1,4 +1,5 @@
 <script lang="ts">
+	import { m } from '$lib/paraglide/messages';
 	import * as Dialog from '&/dialog';
 	import { Label } from '&/label';
 	import LanguageSelect from './language-select.svelte';
@@ -21,20 +22,20 @@
 >
 	<Dialog.Content>
 		<Dialog.Header>
-			<Dialog.Title>Settings</Dialog.Title>
-			<Dialog.Description>Manage pinner settings.</Dialog.Description>
+			<Dialog.Title>{m.noble_east_crab_play()}</Dialog.Title>
+			<Dialog.Description>{m.plain_due_rat_enchant()}</Dialog.Description>
 		</Dialog.Header>
 
 		<div class="flex flex-col items-start gap-5">
 			<Label class="flex w-full flex-col items-start gap-2">
-				Language
+				{m.top_trite_shad_chop()}
 
 				<LanguageSelect />
 			</Label>
 		</div>
 
 		<Dialog.Footer>
-			<Dialog.Close onclick={onClose}>Close</Dialog.Close>
+			<Dialog.Close onclick={onClose}>{m.giant_kind_jay_nail()}</Dialog.Close>
 		</Dialog.Footer>
 	</Dialog.Content>
 </Dialog.Root>