+++ CHANGELOG.md
@@ -4,7 +4,7 @@
 
 ### Changed
 
-- Improved the extension description to better and more clearly communicate its functionality
+- Clarified the extension description to more accurately describe its tab collection management features.
 
 ## [1.1.3] - 2026-09-03