+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Added
 
+- Export collections for Pinner or Save Pinned Tabs
 - Import collections from Pinner or Save Pinned Tabs
 - Add english translation for `SettingsDialog`
 - Add french translation for `SettingsDialog`