+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Added
 
+- Check version mismatch between git tag and `manifest.json` in CI
 - Animate collections accordion items
 - Add appearance mode button in app header
 - Add description section for import/export feature